requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
qdrant-client==1.17.0
sentence-transformers==5.1.0
groq==0.31.0
python-dotenv==1.1.1
